% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_rate.R
\name{get_dark_rate}
\alias{get_dark_rate}
\title{Calculate Dark Rate}
\usage{
get_dark_rate(observed, estimate)
}
\arguments{
\item{observed}{number of observed events}

\item{estimate}{number of estimated events}
}
\value{
Decimal number DRC = 1 - (observed/estimate)
}
\description{
Calculate Dark Rate based on observed and estimated unobserved activity
}
\examples{
DRC <- get_dark_rate(mean(data[,'R']), result_table['C_hat',]$estimates)
DRC <- get_dark_rate(0.59702, 0.7257)
}
