% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_rate.R
\name{get_results}
\alias{get_results}
\title{Combine results}
\usage{
get_results(data, fit, decs = 4)
}
\arguments{
\item{data}{Observed data}

\item{fit}{Return object of optim.R (list of 6)}

\item{decs}{Number of digits to round results}
}
\value{
Dataframe result table with estimates, standard errors and 95% confidence intervals
}
\description{
Combine results after parameter estimation
}
\examples{
parm_sim0 <- set_parms()
parm_sim <- parm_sim0[3:length(parm_sim0)]
data0 <- sim_data(n, parm_sim0)
data <- data0[,1:4]
params0 <- set_inits()
result <- wrap_optim(data, params0, LF)
result_table <- get_results(data, result)
}
