% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_rate.R
\name{LF}
\alias{LF}
\title{Negative Log-Likelihood Function}
\usage{
LF(params, data)
}
\arguments{
\item{params}{Initial parameters for optimization: c(c, phi, eta, d, xi, alpha)}

\item{data}{Observed data: (Z, W, R_prev, R)}
}
\value{
Decimal Value Negative Log-Likelihood
}
\description{
Negative Log-Likelihood for observed and estimated unobserved activity
}
\examples{
loglik <- LF(params, data);
}
