% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_rate.R
\name{LF_gradient}
\alias{LF_gradient}
\title{Gradient Function}
\usage{
LF_gradient(params, data)
}
\arguments{
\item{params}{Initial parameters for optimization: c(c, phi, eta, d, xi, alpha)}

\item{data}{Observed data: (Z, W, R_prev, R)}
}
\value{
Gradient of negative Log-Likelihood at optimization point
}
\description{
Gradient Function of Negative Log-Likelihood Function
}
\examples{
grad <- LF_gradient(params, data);
}
