% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_rate.R
\name{set_inits}
\alias{set_inits}
\title{Initiate Parameters}
\usage{
set_inits(c = 0.5, phi = 0.5, eta = 0.5, d = 0.5, xi = 0.5, alpha = 0.5)
}
\arguments{
\item{alpha}{set initial parameters for optimization: c(c, phi, eta, d, xi, alpha)}
}
\value{
vector of parameters
}
\description{
Initiate Parameters
}
\examples{
params0 <- set_inits();
}
