% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_crime.R
\name{set_parms}
\alias{set_parms}
\title{Set Parameters}
\usage{
set_parms(
  rho = 0,
  tau = 0,
  c = 0.4,
  phi = 0.3,
  eta = -0.2,
  d = 0.5,
  xi = 0.4,
  alpha = 0.2
)
}
\arguments{
\item{alpha}{Parameters for simulation: c(rho, tau, c, phi, eta, d, xi, alpha)}
}
\value{
vector of parameters
}
\description{
Set Parameters
}
\examples{
parm_sim0 <- set_parms();
}
