% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_crime.R
\name{sim_data}
\alias{sim_data}
\title{Simulate Crime}
\usage{
sim_data(n, parms)
}
\arguments{
\item{n}{Number of observations}

\item{parms}{Parameters for simulation: c(rho, tau, c_C, phi, eta, c_E, xi, alpha)}
}
\value{
simulated dataset as matrix
}
\description{
Simulation of entities and time periods with observed and unobserved illegal cases
}
\examples{
data <- sim_data();
}
