% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_rate.R
\name{wrap_optim}
\alias{wrap_optim}
\title{Wrapper Function for optim.R}
\usage{
wrap_optim(data, params, fct, grad = NULL)
}
\arguments{
\item{data}{Observed data}

\item{params}{Initial parameter vector for optimization}

\item{fct}{Objective function}

\item{grad}{Gradient function of objective function (optinal, enhances robustness and runtime of optimization algorithm)}
}
\value{
List of 6: par (estimates), value (optimized value of objective function), counts (number of calls to fn and gr), convergence (0 if converged), message, hessian
}
\description{
Estimates parameter for objective function using optim.R
}
\examples{
parm_sim0 <- set_parms()
parm_sim <- parm_sim0[3:length(parm_sim0)]
data0 <- sim_data(n, parm_sim0)
params0 <- set_inits()
result <- wrap_optim(data[,1:4], params0, LF)
}
